+++ OneKnobVocal/Source/EqualizerEditor.cpp
@@ -63,7 +63,7 @@ void EqualizerEditor::sliderValueChanged(juce::Slider* slider)
             if (abs(slider->getMinValue() - mProcessor.knobValueMap[Equalizer::parameterIDs[i]].start) > 1e-3
                 || abs(slider->getMaxValue() - mProcessor.knobValueMap[Equalizer::parameterIDs[i]].end) > 1e-3)
                 mProcessor.knobValueMap.set(Equalizer::parameterIDs[i],
-                    ModdedNormalisableRange<float>(slider->getMinValue(), slider->getMaxValue()));
+                    ModdedNormalisableRange<double>(slider->getMinValue(), slider->getMaxValue()));
         }
 
     }
