+++ OneKnobVocal/Source/SaturatorEditor.cpp
@@ -62,7 +62,7 @@ void SaturatorEditor::sliderValueChanged(juce::Slider* slider)
             if (abs(slider->getMinValue() - mProcessor.knobValueMap[Saturator::parameterIDs[i]].start) > 1e-3
                 || abs(slider->getMaxValue() - mProcessor.knobValueMap[Saturator::parameterIDs[i]].end) > 1e-3)
                 mProcessor.knobValueMap.set(Saturator::parameterIDs[i],
-                    ModdedNormalisableRange<float>(slider->getMinValue(), slider->getMaxValue()));
+                    ModdedNormalisableRange<double>(slider->getMinValue(), slider->getMaxValue()));
         }
 
     }
